requirements.txt
Flask>=2.0
python-whois
qrcode[pil]
validators
tldextract
